refactor: :truck: moved `examples_*()` into `zenodo.py`, circular imports
