+++ src/zen_do/zenodo.py
@@ -394,3 +394,54 @@ class ZenodoClient:
             timeout=self.timeout,
         )
         return self._resolve(response, ZenodoRecord)
+
+
+def example_metadata(title: str = "Test Book") -> ZenodoMetadata:
+    """A set of example Zenodo metadata."""
+    return ZenodoMetadata(
+        title=title,
+        upload_type="poster",
+        creators=[
+            ZenodoCreator(
+                name="Doe, Jane", affiliation="University of Testfalia", orcid="ABC"
+            )
+        ],
+        related_identifiers=[
+            ZenodoRelatedIdentifier(
+                identifier="urn:zenodo:my-org:project:book",
+                relation="isIdenticalTo",
+                resource_type="other",
+                scheme="urn",
+            )
+        ],
+    )
+
+
+def example_record(
+    id: int = 123,
+    metadata: ZenodoMetadata = example_metadata(),
+    state: ZenodoRecordState = "done",
+    submitted: bool = True,
+    bucket: Optional[str] = "https://path.com/path/wrwee-324-23f-sdf",
+    urn: str = "urn:zenodo:my-org:project:book",
+) -> ZenodoRecord:
+    """An example Zenodo record."""
+    metadata = metadata.model_copy(
+        update={
+            "related_identifiers": [
+                ZenodoRelatedIdentifier(
+                    identifier=urn,
+                    relation="isIdenticalTo",
+                    resource_type="other",
+                    scheme="urn",
+                )
+            ]
+        }
+    )
+    return ZenodoRecord(
+        id=id,
+        metadata=metadata,
+        state=state,
+        submitted=submitted,
+        links=ZenodoLinks(latest_draft="https://path.com/path/9999", bucket=bucket),
+    )
